More reorganization, override syntax
